pyproject.toml
[project]
name = "transcribe"
version = "0.1.0"
description = "Real-time microphone transcription using Mistral's Voxtral API"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "mistralai[realtime]",
    "sounddevice",
    "numpy",
    "python-dotenv",
]
